+++ R/discretize.R
@@ -62,6 +62,10 @@ discretize <- function(x, method = "frequency", breaks = 3,
 
 
 discretizeDF <- function(df, methods = NULL, default = NULL) {
+  
+  ### methods is a data.frame to get the discretization info from
+  if(is.data.frame(methods)) return(.rediscretizeDF(methods, df))
+  
   for(i in colnames(df)) {
     if(is.logical(df[[i]])) next
     if(is.numeric(df[[i]])) {
@@ -69,9 +73,20 @@ discretizeDF <- function(df, methods = NULL, default = NULL) {
       if(!is.null(methods[[i]])) args <- methods[[i]]
       df[[i]] <- do.call("discretize", c(list(x = df[[i]]), args))
     }
-    
-    if(!is.factor(df[[i]])) df[[i]] <- as.factor(df[[i]])
   }
   
   df
+}
+
+
+.rediscretizeDF <- function(data, newdata) {
+  
+  if(!all(colnames(data) == colnames(newdata))) stop("columns in data and newdata do not conform!")
+  
+  cps <- lapply(data, FUN = function(x) { 
+    list(breaks = attr(x, "discretized:breaks"), method = "fixed", 
+      labels = levels(x))
+    })
+  
+  discretizeDF(newdata, methods = cps)
 }