+++ NEWS.md
@@ -1,5 +1,6 @@
 # arules 1.5-5.1 (xx/xx/2018)
 
+* Fixed typo in inspect for tidLists (reported by Carlos Chavarria). 
 
 # arules 1.5-5 (01/09/2018)