+++ NEWS.md
@@ -1,7 +1,11 @@
 # arules 1.5-5.1 (xx/xx/2018)
 
-* Fixed typo in inspect for tidLists (reported by Carlos Chavarria). 
+## New Features
 * Added interest measure "importance".
+* Added method items for transactions.
+
+## Bug Fixes
+* Fixed typo in inspect for tidLists (reported by Carlos Chavarria). 
 
 # arules 1.5-5 (01/09/2018)